+++ tests/testthat/test_5_Tasks.R
@@ -17,6 +17,7 @@ test_that("textClassify tests", {
     function_to_apply = "none"
   )
   expect_equal(sen1$score_x, 4.67502, tolerance = 0.001)
+  textModelsRemove("distilbert-base-uncased-finetuned-sst-2-english")
 
   #  # Test another model
   sen2 <- textClassify("I like you. I love you",
@@ -25,6 +26,7 @@ test_that("textClassify tests", {
     return_all_scores = TRUE,
     function_to_apply = "none"
   )
+  textModelsRemove("cardiffnlp/twitter-roberta-base-sentiment")
   #
 })
 
@@ -45,7 +47,6 @@ test_that("textGeneration test", {
     prefix = "",
     handle_long_generation = "hole"
   )
-
   expect_equal(generated_text$x_generated, "The meaning of life is self-improvement. We do this the right way. It is an exercise in self-improvement. The better part of the day we try to do that is to stop thinking about doing things to the level most people")
   expect_that(generated_text$x_generated, is_a("character"))
 
@@ -65,6 +66,7 @@ test_that("textGeneration test", {
     handle_long_generation = "hole",
     set_seed = 22L
   )
+  textModelsRemove("gpt2")
   expect_equal(generated_text2$generated_token_ids[1], 464)
   expect_that(generated_text2$generated_token_ids[1], is_a("integer"))
 })
@@ -88,6 +90,7 @@ test_that("textNER test", {
   )
   ner_example2
   expect_equal(ner_example2$satisfactiontexts_NER$score[2], 0.976, tolerance = 0.01)
+  textModelsRemove("dslim/bert-base-NER")
 })
 
 test_that("textSum test", {
@@ -121,22 +124,28 @@ test_that("textZeroShot test", {
   ZeroShot_example <- text::textZeroShot(
     sequences = c("I play football", "The forrest is wonderful"),
     candidate_labels = c("sport", "nature", "research"),
-    model = "facebook/bart-large-mnli"
+    #model = "facebook/bart-large-mnli"
+    model = "okho0653/distilbert-base-uncased-zero-shot-sentiment-model"
   )
 
-  expect_equal(ZeroShot_example$scores_x_1[1], 0.9985854)
+  testthat::expect_equal(ZeroShot_example$scores_x_1[1], 0.3341856, tolerance = 0.00001)
+  textModelsRemove("okho0653/distilbert-base-uncased-zero-shot-sentiment-model")
 })
 
 test_that("textTranslate test", {
   skip_on_cran()
 
+  textModels()
   translation_example <- text::textTranslate(
     Language_based_assessment_data_8[1, 1:2],
     source_lang = "en",
     target_lang = "fr",
-    model = "t5-base"
+    model = "t5-small",
+    max_length = 400
   )
-  translation_example
-  expect_that(translation_example$en_to_fr_satisfactiontexts, is_a("character"))
-  expect_equal(translation_example$en_to_fr_satisfactiontexts[1], "Je ne suis pas satisfait de ma vie. Je suis reconnaissante de ce que j'ai et de l'endroit où je suis parce que la situation peut toujours être pire. Je veux une carrière et un diplôme, je veux perdre du poids et je n'ai pas encore atteint ces objectifs. Je ne suis donc pas satisfait de ma vie à l'heure actuelle.")
+
+  testthat::expect_that(translation_example$en_to_fr_satisfactiontexts, testthat::is_a("character"))
+  testthat::expect_equal(translation_example$en_to_fr_satisfactiontexts[1],
+                         "Je ne suis pas satisfait de ma vie, je suis reconnaissante de ce que j'ai et de ce que je suis, car la situation peut toujours être pire. Je veux une carrière et un diplôme, je veux perdre de poids et je n'ai pas encore atteint ces objectifs.")
+  textModelsRemove("t5-small")
 })